new file:   ModuleBase/product/moddef/CProductDef.php

+++ ModuleBase/user/action/dep_edit.php
@@ -16,8 +16,10 @@ if(isset($_REQUEST['id'])){
 		$udep->setPrimaryKey($id);
 		$udepmbr->setPrimaryKey($id);
 		if(isset($_REQUEST['delete'])){
-			$ret = $udep->destroy();
-			$ret = $udepmbr->destroy();
+			if($id > CUserDef::BANNED_DEL_MAX_CLASS_ID){
+				$ret = $udep->destroy();
+				$ret = $udepmbr->destroy();
+			}
 		}
 		else if(isset($_REQUEST['edit_submit'])){
 			$edit_info = array(
@@ -108,7 +110,7 @@ if(isset($_REQUEST['id'])){
 	<?php }?>
 	<div <?php echo 1==$k%2 ? 'class="even"' : ''?>>
 	<div class="inpBox mb17">
-   		<label for="name" class="labelL"><?php echo $mbs_appenv->lang('name')?>&nbsp;:&nbsp;</label>
+   		<label for="name" class="labelL"><?php echo $mbs_appenv->lang('name', 'common')?>&nbsp;:&nbsp;</label>
 	    <input id="name" class="inpTit" name="name[]" type="text" value="<?php echo $row['name']?>" 
 	    	placeholder="<?php echo $mbs_appenv->lang('please_input')?>" required />
     </div>
